reference/scenarios/litellm/pyproject.toml
[project]
name = "litellm-reference-test"
version = "0.0.0"
requires-python = ">=3.12"
dependencies = [
    "litellm==1.83.4",
    "genai-reference-shared",
]

[tool.uv.sources]
genai-reference-shared = { path = "../../shared", editable = true }

[tool.uv]
package = false
